requirements.txt
pywebview>=4.4.1
anthropic>=0.25.0
